Rewrote field class to integrate Meltdown.
This changes the field to extend the standard cfs_field instead of
cfs_textarea. There are some pretty significant differences between the
two fields now, so it makes more sense to be a unique field type. This
also fixed the class file name which was a typo and should have been
named this to begin with.